move bump and abandon test to serviceTest
