+++ src/test/java/bio/terra/janitor/app/controller/JanitorApiControllerTest.java
@@ -218,319 +218,4 @@ public class JanitorApiControllerTest {
             .collect(Collectors.toList()),
         Matchers.containsInAnyOrder(id1, id2));
   }
-
-  @Test
-  public void abandoResources() throws Exception {
-    CloudResourceUid resourceUid =
-        new CloudResourceUid()
-            .googleProjectUid(new GoogleProjectUid().projectId(UUID.randomUUID().toString()));
-    CreateResourceRequestBody body1 =
-        new CreateResourceRequestBody()
-            .resourceUid(resourceUid)
-            .creation(CREATION)
-            .expiration(EXPIRATION)
-            .labels(DEFAULT_LABELS);
-    CreateResourceRequestBody body2 =
-        new CreateResourceRequestBody()
-            .resourceUid(resourceUid)
-            .creation(CREATION)
-            .expiration(EXPIRATION.plusMinutes(1))
-            .labels(DEFAULT_LABELS);
-
-    String createResponse1 =
-        this.mvc
-            .perform(
-                post("/api/janitor/v1/resource")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(body1))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    String createResponse2 =
-        this.mvc
-            .perform(
-                post("/api/janitor/v1/resource")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(body2))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    this.mvc
-        .perform(
-            put("/api/janitor/v1/resource")
-                .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                .queryParam("state", ResourceState.ABANDONED.toString())
-                .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-        .andDo(MockMvcResultHandlers.print())
-        .andExpect(status().isNoContent());
-
-    String getResponse =
-        this.mvc
-            .perform(
-                get("/api/janitor/v1/resource")
-                    .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    String olderId = objectMapper.readValue(createResponse1, CreatedResource.class).getId();
-    String newerId = objectMapper.readValue(createResponse2, CreatedResource.class).getId();
-
-    TrackedResourceInfoList resourceInfoList =
-        objectMapper.readValue(getResponse, TrackedResourceInfoList.class);
-    TrackedResourceInfo olderResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(olderId))
-            .findFirst()
-            .get();
-    TrackedResourceInfo newerResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(newerId))
-            .findFirst()
-            .get();
-    assertEquals(2, resourceInfoList.getResources().size());
-    assertEquals(TrackedResourceState.DUPLICATED.name(), olderResource.getState());
-    assertEquals(TrackedResourceState.ABANDONED.name(), newerResource.getState());
-
-    // Then bump the abandoned resource
-    this.mvc
-        .perform(
-            put("/api/janitor/v1/resource")
-                .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                .queryParam("state", ResourceState.READY.toString())
-                .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-        .andDo(MockMvcResultHandlers.print())
-        .andExpect(status().isNoContent());
-
-    getResponse =
-        this.mvc
-            .perform(
-                get("/api/janitor/v1/resource")
-                    .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    resourceInfoList = objectMapper.readValue(getResponse, TrackedResourceInfoList.class);
-    olderResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(olderId))
-            .findFirst()
-            .get();
-    newerResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(newerId))
-            .findFirst()
-            .get();
-    assertEquals(2, resourceInfoList.getResources().size());
-    assertEquals(TrackedResourceState.DUPLICATED.name(), olderResource.getState());
-    assertEquals(TrackedResourceState.READY.name(), newerResource.getState());
-  }
-
-  @Test
-  public void bumpResources() throws Exception {
-    CloudResourceUid resourceUid =
-        new CloudResourceUid()
-            .googleProjectUid(new GoogleProjectUid().projectId(UUID.randomUUID().toString()));
-    CreateResourceRequestBody body1 =
-        new CreateResourceRequestBody()
-            .resourceUid(resourceUid)
-            .creation(CREATION)
-            .expiration(EXPIRATION)
-            .labels(DEFAULT_LABELS);
-    CreateResourceRequestBody body2 =
-        new CreateResourceRequestBody()
-            .resourceUid(resourceUid)
-            .creation(CREATION)
-            .expiration(EXPIRATION.plusMinutes(1))
-            .labels(DEFAULT_LABELS);
-
-    String createResponse1 =
-        this.mvc
-            .perform(
-                post("/api/janitor/v1/resource")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(body1))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    String createResponse2 =
-        this.mvc
-            .perform(
-                post("/api/janitor/v1/resource")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(body2))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    this.mvc
-        .perform(
-            put("/api/janitor/v1/resource")
-                .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                .queryParam("state", ResourceState.ABANDONED.toString())
-                .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-        .andDo(MockMvcResultHandlers.print())
-        .andExpect(status().isNoContent());
-
-    String getResponse =
-        this.mvc
-            .perform(
-                get("/api/janitor/v1/resource")
-                    .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    String olderId = objectMapper.readValue(createResponse1, CreatedResource.class).getId();
-    String newerId = objectMapper.readValue(createResponse2, CreatedResource.class).getId();
-
-    TrackedResourceInfoList resourceInfoList =
-        objectMapper.readValue(getResponse, TrackedResourceInfoList.class);
-    TrackedResourceInfo newerResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(newerId))
-            .findFirst()
-            .get();
-    assertEquals(TrackedResourceState.ABANDONED.name(), newerResource.getState());
-
-    // Then bump the abandoned resource
-    this.mvc
-        .perform(
-            put("/api/janitor/v1/resource")
-                .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                .queryParam("state", ResourceState.READY.toString())
-                .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-        .andDo(MockMvcResultHandlers.print())
-        .andExpect(status().isNoContent());
-
-    getResponse =
-        this.mvc
-            .perform(
-                get("/api/janitor/v1/resource")
-                    .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    resourceInfoList = objectMapper.readValue(getResponse, TrackedResourceInfoList.class);
-    TrackedResourceInfo olderResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(olderId))
-            .findFirst()
-            .get();
-    newerResource =
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(newerId))
-            .findFirst()
-            .get();
-    assertEquals(2, resourceInfoList.getResources().size());
-    assertEquals(TrackedResourceState.DUPLICATED.name(), olderResource.getState());
-    assertEquals(TrackedResourceState.READY.name(), newerResource.getState());
-  }
-
-  @Test
-  public void abandonResources_notFound() throws Exception {
-    CloudResourceUid resourceUid =
-        new CloudResourceUid()
-            .googleProjectUid(new GoogleProjectUid().projectId(UUID.randomUUID().toString()));
-    this.mvc
-        .perform(
-            put("/api/janitor/v1/resource")
-                .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                .queryParam("state", ResourceState.ABANDONED.toString())
-                .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-        .andDo(MockMvcResultHandlers.print())
-        .andExpect(status().isNotFound());
-  }
-
-  /** Gets 404 error when resource exists but not in ABANDONED state. */
-  @Test
-  public void bumpResources_notFound() throws Exception {
-    CloudResourceUid resourceUid =
-        new CloudResourceUid()
-            .googleProjectUid(new GoogleProjectUid().projectId(UUID.randomUUID().toString()));
-    CreateResourceRequestBody body =
-        new CreateResourceRequestBody()
-            .resourceUid(resourceUid)
-            .creation(CREATION)
-            .expiration(EXPIRATION)
-            .labels(DEFAULT_LABELS);
-
-    String createResponse =
-        this.mvc
-            .perform(
-                post("/api/janitor/v1/resource")
-                    .contentType(MediaType.APPLICATION_JSON)
-                    .content(objectMapper.writeValueAsString(body))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    String id = objectMapper.readValue(createResponse, CreatedResource.class).getId();
-
-    // Bump the resource
-    this.mvc
-        .perform(
-            put("/api/janitor/v1/resource")
-                .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                .queryParam("state", ResourceState.READY.toString())
-                .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-        .andDo(MockMvcResultHandlers.print())
-        .andExpect(status().isNotFound());
-
-    // Verify the resource is READY
-    String getResponse =
-        this.mvc
-            .perform(
-                get("/api/janitor/v1/resource")
-                    .queryParam("cloudResourceUid", objectMapper.writeValueAsString(resourceUid))
-                    .header(AuthHeaderKeys.OIDC_CLAIM_EMAIL.getKeyName(), ADMIN_USER_EMAIL))
-            .andDo(MockMvcResultHandlers.print())
-            .andExpect(status().isOk())
-            .andReturn()
-            .getResponse()
-            .getContentAsString();
-
-    TrackedResourceInfoList resourceInfoList =
-        objectMapper.readValue(getResponse, TrackedResourceInfoList.class);
-    assertEquals(1, resourceInfoList.getResources().size());
-    assertEquals(
-        TrackedResourceState.READY.name(),
-        resourceInfoList.getResources().stream()
-            .filter(resource -> resource.getId().equals(id))
-            .findFirst()
-            .get()
-            .getState());
-  }
 }
